build: Link YARA statically for Linux
Using yari as a library fails to link libyara in the parent crate. This
is because the link-search-path is not propagated to parent crates
(because it does not follow all requirements).

There are many paths we can follow, we will need to adopt a more
standard way in the future to handle linking libs.